api/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
slowapi>=0.1.9

# Database
supabase>=2.3.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OCR and Image Processing
pytesseract==0.3.10
Pillow>=10.1.0
pdf2image==1.16.3

# ML and Embeddings (for item mapping)
sentence-transformers>=2.2.2
numpy>=1.26.2

# Task Queue
celery==5.3.4
redis==5.0.1

# Object Storage
boto3==1.34.14
minio==7.2.3

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Linting and Formatting
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.7.1
bandit==1.7.6
